requirements.txt
# Django framework
Django==5.0.6

# Database
psycopg2-binary==2.9.9

# REST API
djangorestframework==3.15.1

# Environment variables
python-decouple==3.8

# Email functionality
django-crispy-forms==2.1

# PDF generation
reportlab==4.0.7

# Image processing
Pillow==10.3.0

# Caching
redis==5.0.3

# Development tools
django-debug-toolbar==4.4.2
